+++ src/components/Docs/types/types.ts
@@ -1,0 +1,3 @@
+export function isNonNull<T>(value: T): boolean {
+  return value !== null && value !== undefined;
+}
